Move `<group-list>` component to src/sidebar/components
