+++ src/sidebar/components/group-list.js
