+++ src/sidebar/components/test/group-list-test.js
@@ -3,7 +3,7 @@
 var angular = require('angular');
 
 var groupList = require('../group-list');
-var util = require('./util');
+var util = require('../../directive/test/util');
 
 describe('groupList', function () {
   var $window;
